requirements.txt
einops
einops_exts
flow_vis
imageio
json_tricks
matplotlib
numpy<2.0
opencv_contrib_python
opencv_python
pandas
Pillow
PyYAML
Requests
rotary_embedding_torch
scikit_image
scikit_learn
scipy
torch==2.2.0+cu118
torchvision==0.17.0+cu118
torchaudio==2.2.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
tqdm
transformers==4.41.0
accelerate==0.24.1
pytorch-fid
torchmetrics
torchmetrics[image]
openai-clip